pyproject.toml
[project]
name = "rooki_ai"
version = "0.1.0"
description = "rooki-ai using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=0.165.1,<1.0.0",
    "fastapi[standard]>=0.109.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.0.0",
    "httpx>=0.24.0",
    "emoji>=2.0.0",
    "psycopg2-binary>=2.9.10",
    "asyncpg>=0.30.0",
]

[project.scripts]
rooki_ai = "rooki_ai.main:run"
run_crew = "rooki_ai.main:run"
train = "rooki_ai.main:train"
replay = "rooki_ai.main:replay"
test = "rooki_ai.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
